Be consistent with including cryptoConfig.h.
Every other file uses just '#include "cryptoConfig.h"',
and uses an -I to the MATRIXSSL_ROOT.

Because also in -I exists to MATRIXSSL_ROOT/crypto,
including "../cryptoConfig.h" works as long as
MATRIXSSL_ROOT points to the root of the source tree.

However, using automake I will be generating cryptoConfig.h
in the root of the build directory, so I'll be using
a -I$(builddir) and "../cryptoConfig.h" doesn't work
anymore.